refactor: extract raceWithSignal
