+++ src/abort.ts
@@ -4,27 +4,7 @@ import { isPromiseLike } from '.'
 import is from '@sindresorhus/is'
 import { AbortError } from './AbortError'
 import { bubbleAbort } from './bubbleAbort'
-import { composeAbort } from './composeAbort'
-
-export async function raceWithSignal <TReturn = unknown> (command: (signal: AbortSignal) => Iterable<Promise<TReturn>>, inputSignal?: AbortSignal): Promise<TReturn> {
-  const { signal, abort } = composeAbort(inputSignal)
-  const promises = Array.from(command(signal))
-  if (!is.nullOrUndefined(inputSignal)) {
-    promises.push(new Promise((resolve, reject) => {
-      const abortHandler = (): void => {
-        clear()
-        reject(new AbortError())
-      }
-      inputSignal.addEventListener('abort', abortHandler)
-      const clear = (): void => {
-        inputSignal.removeEventListener('abort', abortHandler)
-        signal.removeEventListener('abort', clear)
-      }
-      signal.addEventListener('abort', clear)
-    }))
-  }
-  return await Promise.race(promises).finally(abort)
-}
+import { raceWithSignal } from './raceWithSignal'
 
 const noop = (): void => {}
