+++ src/index.ts
@@ -5,5 +5,6 @@ export * from './AbortError'
 export * from './bubbleAbort'
 export * from './extPromise'
 export * from './isPromiseLike'
+export * from './raceWithSignal'
 export * from './toPromise'
 export * from './types'
